Separation of concerns: extracted config variable to its own file and out of main file
