+++ src/assets/license.txt
@@ -1,0 +1,5 @@
+All assets used in this tutorial are CC0 licensed. 
+The tiles are created by Kenney Vleugels and can be found at www.kenney.nl 
+
+The player sprites can be found here .
+https://opengameart.org/content/rpg-character-sprites
